+++ src/smart_core_assistant_painel/modules/initial_loading/features/firebase_init/domain/usecase/firebase_init_usecase.py
@@ -27,7 +27,6 @@ class FirebaseInitUseCase(FIUsecase):
                 firebase_admin.get_app()
             except ValueError:
                 firebase_admin.initialize_app()
-            print("Firebase Admin SDK inicializado com sucesso!")
 
             return SuccessReturn(EMPTY)