+++ src/smart_core_assistant_painel/modules/services/utils/types.py
@@ -1,0 +1,14 @@
+from typing import TypeAlias
+
+from py_return_success_or_error import Datasource, Empty, UsecaseBaseCallData
+
+from smart_core_assistant_painel.modules.services.utils.parameters import (
+    SetEnvironRemoteParameters,
+)
+
+SERUsecase: TypeAlias = UsecaseBaseCallData[
+    Empty,
+    bool,
+    SetEnvironRemoteParameters,
+]
+SERData: TypeAlias = Datasource[bool, SetEnvironRemoteParameters]